Solution for rank
Part 2 of the implementation of an AVL Tree - obtaining rank of each node, with part 1 being under the avl folder of ADT Challenge. Given N elements in an AVL Tree, 1 is the rank of the smallest element while N is the largest. BST.java contains the implementation while Grader.java contains the client class for input handling and testing purposes.